Move hooks to own folder
