+++ client/app/pages/queries/hooks/useDataSourceSchema.js
