+++ client/app/pages/queries/hooks/useVisualizationTabHandler.js
